Rename : splash -> intro
