+++ core/navigation/build.gradle.kts
@@ -14,7 +14,7 @@ dependencies {
     implementation(projects.feature.login.api)
     implementation(projects.feature.grouphome.api)
     implementation(projects.feature.segmentation.api)
-    implementation(projects.feature.splash.api)
+    implementation(projects.feature.intro.api)
     implementation(projects.feature.groupenter.api)
 
     implementation(libs.bundles.navigation)
