+++ core/navigation/src/main/java/com/teamyg/parfait/core/navigation/NavigatorConst.kt
@@ -1,6 +1,6 @@
 package com.teamyg.parfait.core.navigation
 
-import com.teamyg.parfait.feature.splash.api.NavKeySplash
+import com.teamyg.parfait.feature.intro.api.NavKeySplash
 
 object NavigatorConst {
     val INITIAL_NAVIGATION_KEY = NavKeySplash
