+++ feature/intro/impl/src/main/java/com/teamyg/parfait/feature/intro/impl/EntryBuilder.kt
@@ -1,4 +1,4 @@
-package com.teamyg.parfait.feature.splash.impl
+package com.teamyg.parfait.feature.intro.impl
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
@@ -6,7 +6,7 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.ui.Modifier
 import androidx.navigation3.runtime.EntryProviderScope
 import androidx.navigation3.runtime.NavKey
-import com.teamyg.parfait.feature.splash.api.NavKeySplash
+import com.teamyg.parfait.feature.intro.api.NavKeySplash
 import com.teamyg.parfait.core.navigation.Navigator
 
 fun EntryProviderScope<NavKey>.featureSplashEntryBuilder(navigator: Navigator) {
