+++ feature/intro/impl/src/main/java/com/teamyg/parfait/feature/intro/impl/SplashScreen.kt
@@ -1,4 +1,4 @@
-package com.teamyg.parfait.feature.splash.impl
+package com.teamyg.parfait.feature.intro.impl
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
